add custom classpath in run script and text output
The default runscript does not point to the correct file path. There is
also a bug which stores multiple lines in one sentence as various keys.

This causes the run script to fail and the bug causes more items in the
list

Modifying the class path and bin path in the run script will allow
compilation and succesful running. No compilation error with regards to
packages

Let's make the class path to run the main function in Duke point to bin
and the package has the remaining path