Duke class: change format of greeting message
The greeting message does not comply with the intended format for all
messages produced by the bot.

Let's create a method printMessage(String) that will take in an
unformatted message and print it out in the bot's message format. The
greeting message is then printed using this method call.

The bot is expected to print other messages that must also comply with
the message format. Using a method that will handle the formatting will
reduce code duplication.