Duke class: change to support user input with echo and bye command
User's need to be able to input echo and bye commands.

Let's change main(String[]) to have a loop that takes user inputs and
for each input, run the appropriate command and display the resulting
message. Exit the loop when the bye command is inputted.

The code is written such that future modifications can be easily made to
account for other command types.