Add support for add command
The user needs to add tasks to their task list.

Let's add an AddCommand class which implements the Command interface and
have the command run whenever the user inputs anything other than "bye".

An AddCommand class would encapsulate the logic for the add command type
and can be utilised through the Command interface.