Add support for list command
The user needs to list out all tasks in their task list.

Let's add a ListCommand class which implements the Command interface and
have the command run whenever the user inputs "list".

A ListCommand class would encapsulate the logic for the list command
type and can be utilised through the Command interface.