Allow Babe to detect instructions more efficiently
Prior to this, in order to detect a status change for an item,
Babe will extract the first word of userInput string in a separate
process and check for for "mark" and "unmark".

However, this requires code that caters to each instruction such as
"mark" and "unmark" to ensure that extraction of substring does not
encounter NullPointerException.

By making userInput an ArrayList of String type, and parsing the
string into the ArrayList, to obtain an instruction from a user
input (apart from add item functionality), we simply have to get
the first element from the ArrayList.

Thus, in this commit, we:
- Change userInput from type String to ArrayList
- Remove checkChangeStatus() that detects whether current
instruction asks for mark/unmark
- Add helper function to rebuild user input from the ArrayList