Create Item class for all list items in Babe
In Level-4 (yet to be implemented), the list items in Babe now has a
few different types, namely, ToDos, Deadlines and Events.

Each of these Items keeps track of different data, and has different
String representations.

Thus, in this commit, Item is abstracted from Babe's list as a
standalone abstract class, with each different type of Item as its
subclasses for ease of usage and better encapsulation.

Each Item now stores a String content and additional String data
such as deadlines and start/end date depending on the object. The
Done status is also now recorded in each Item.