Closes #5, creates and implements DukeException
Errors in user input are not caught and runtime exceptions are thrown.

This not only causes Duke to exit ungracefully but also causes the
unsmooth operation of duke.

Thus, I create a DukeException class, extending the Exception
class, which can be thrown with an error message.

Having our own checked exception allows our own exception to be thrown,
caught and then handled to tell the user what was wrong.

This is also done to conform to Level-5
https://nus-cs2103-ay2223s2.github.io/website/schedule/week2/project.html