Replace List<Task> with TaskList
There are operations that a task list should have which does not exist
in List such as a toString that is specific to TaskList.

Let's create a TaskList class and replace all usage of List<Task> with
TaskList.

This allows us to add task list specific operations as methods to
TaskList.