Added Assertion Feature
The previous version of Mr. Bear Does not have any checking mechanisms
to verify the format of the event stored in the storage file as I
assumed that events would be written to the file in the correct format.
However, this cannot be guaranteed and should any event be stored in
the wrong format and loaded incorrectly, the whole program would
potentially crash. Therefore, it is the necessary to add an
assertion feature to verify the assumption is indeed true.

Add assertion to check whether the event type are of the format [T],
[D], [E], representing TodoTask, DeadlineTask, and EventTask
respectively. Add assertion to check whether the isDone status is of
the format [ ], or [X]. If not, terminate the program and display the
relevant error message.