Add automatic duke_data save
So that the data will not only be in memory but also a txt file.

Currently it is at current working directory, the files will be saved
at different places when running in different enviroments.

Also ignored the duke_data.txt in gitignore, user data not stored in VC

TODO: Figure a better place to store the file.