Implements serialization to task, storage class to handle files
There is no way to preserve sessions. Users have to retype everything
which is a pain.

I create a storage class to deal with files, both storing and loading.
Writing to a file and reading from a file becomes easier because of
serialization.

This should make saving and loading possible and easier to implement.