Added Help Feature
The current version of Mr. Bear does not have an in-app help menu that allows user to check the format of the their command. Although it does have error prompt to notify the user when the input invalid command, it is still rather inconvenient.

Therefore, I added an help command to allow the user to check all the available command and their respective functionalities by displaying it in a dialog box.