Add Save feature
Duke is running on data that is not saved and resets every time the
Duke program is rerun.

Tasks should be saved and used for future rerun of Duke for the users.

Let's add a folder with a text file and call Scanner class to read the
text file and FileWriter class to write in the text file.

Using Scanner class to read the text file and FileWriter class to write
 in the text file helps to store task in a text file as expected.