Add dates formatting for deadline
The toString method in Deadline class prints the deadline in a String
format provided by the user.

The toString method should print in the mmm-dd-yyyy format instead of a
string.

Let's import the LocalDate class and parse in the date to the var by.
Use the var by of type LocalDate to print the date in mmm-dd-yyyy format
.

LocalDate class helps to provide a format that can be printed out for
the user.