Implement loading function
When the bot starts up, it should read the save file, if any, and load
the state of the database.

KiraBot now reads the file and load in the saved state of the task list.
Further changes are made to change the format of the delimiters to allow
for commas in task description.