Refactor code to include more OOP and packages
Currently, there is no organisation for the different java files and no
dedicated class for each logic abstraction.

More classes have been implemented to separate the logic and taskList no
longer return formatted strings for each method as it is not the job.
Furthermore, Parser will manage all of the error during parsing of the
commands.