Add auto-saving feature for task history
Creation of file to store task history needs to be coupled with saving
of task history when the user changes the task list.

Let's add a feature that writes the contents of the user's task list
to the data file whenever the user makes changes to the list.

In the event that the program abruptedly terminated, prior changes to
the task list before termination would still have been saved.