Added Update Task Description Feature
In order to update the description of a task in the current version of Mr. Bear, the user has no choice but to delete the task first and added a new one with the intended description, which is very tedious.

Therefore, a direct update function is added to allow the users to alter the description easily and the help menu is also changed.