Add a help command
With a greater number of commands, it is difficult to remember all
the available commands. It is worse for new users of the chatbot.

Add a help command that will display the list of available commands.

User can now easily view all available commands if he is unsure of what
to type into the chatbot.