Add command to check tasks on a specific day.
Provide users a way to focus on important tasks and narrow down
the search.

Let's add a on date command. Will display all tasks that occur on the
given date. Loop through the ArrayList and process the dates.

Have to check what the type of Task is, and carry out date processing
according to the correct type.