Refactoring of code to make it more OOP
Current situation: instances of Storage and Tasklist had Ui in its constructor.

TaskList had both a Storage and Ui, and Sotrage had Ui again.

This was not ideal so this was removed as it did not adhere to OOP principle.

Code was refactored and methods utilising Ui was changed to continue to function as per normal.