feat: Save/ Load feature implemented.
Duke can now save each session to a .txt file and load from that file in the future sessions. Save takes in the ArrayList of Tasks and serializes through the Serializable interface. The byte stream is then written to 'duke.txt' file. The object is then retrieved in the future session. This is immediately assigned to the new session's ArrayList of Tasks.

If directory to duke.txt does not exist, it is created. If the directory and file exists, it is overwritten to the latest session of save.