Add more classes for OOP
Program is cluttered in static main with little OOP.

Program code looks untidy with no OOP in mind.

Let's add more OOP classes such as Parser to deal with making sense of
the user command, Storage to deal with loading tasks from the file and
saving tasks in the file, TaskList to contain the task list and UI to
deal with user interactions.

These classes can help to further categorize the different objects in
an OOP way which makes the code more readable.