Add JUnit tests for classes in task package and move some classes into extra folder
Classes are not in the right folder. No JUnit tests currently available.

JUnit tests are required to test the different classes. Classes are not
in right folder will affect the test.

Let's create a seedu folder for easier navigation and create JUnit tests
classes for the classes in the task package.

JUnit tests allow semi-automation of test for the respective classes.