Add Find feature
No find feature for the user to find what tasks is in the TaskList.

User must be able to find their tasks if it is done or not.

Let's add some find methods in TaskList and Ui to support this issue.

TaskList helps to find the matching tasks and Ui helps to display the
tasks and messages.