Added find functionality
Current situation: there is no way to find a keyword in the list.

Find functionality is a very helpful function to use for this bot.

Added find functionality in TaskList to print every single tasks that has the keyword in their taskname.

This meant that finding "work" will find "homework", this is an intended effect.