Add Dates Feature
Duke can store the deadline dates as strings for a deadline task.

To make the dates more meaningful, the dates should be stored as dates.

The dates are passed as strings, then it is parsed as date (using
java's LocalDate).

Input from user must be in the format of yyyy-mm-dd. Otherwise, it will
be invalid. Times of the deadline are not yet included in this feature.