Fix bugs for Level-8 combined with Level-7
While getting the data from the saved file, a bug will occur if the
type of task is a deadline. This is because the date format in the
saved file is MM dd yyyy instead of yyyy-mm-dd. This will cause the
constructor of the Deadline class to accept MM dd yyyy instead of
yyyy-mm-dd which will cause an error in the constructor.

To fix the error, we should change the date format in the constructor
that is being used to obtain the data from the saved file.