Improve OOP usage and organization
All classes are in the same directory and the main duke file is very
long as a lot of logic that can be abstracted to other classes are
in there.

Better OOP practices need to be followed and the main duke class
needs to be more readable and less cluttered.

Lets,
* Encapsulate logic behind parsing user input and printing to output
into the Ui and Parser classes
* Encapsulate all commands using Command classes and move away from
simple enumerations
* Package and refactor files into better organized directories.
* Update the automated testing scripts