Reconstruct files to make the code more OOP
The logic of the application is in the Duke.java file, so this does not
follow the OOP Principles: abstraction, encapsulation, polymorphism,
and inheritance.

To follow those principles, we must make separate files and designate
each class to have a specific function (e.g. Parse -> to parse the
user's commands, Ui -> to deal with user interactions, Command -> to
execute the command that the user enters, etc.)

The new classes made are: Ui, Parse, Storage, TaskList, Command
(abstract class), AddCommand, DeleteCommand, ListCommand, MarkCommand,
UnmarkCommand, ExitCommand, CommandType (enum class)

Each class have different functions and the command types are separated
into different classes so as to make the code clear and provide
encapsulation.

Used the enum class for CommandType because there is a finite number
of commands that a user can enter (which makes it suitable according
to the textbook)