Add basic file support to open /data/data.txt
Need a system to save and load tasks for Duke.

Java NIO Paths class is used to load up directory for Duke's data
and load data.txt file to load any existing tasks from previous
execution of the program.

Structure for data.txt on how it will be saved is implemented later.