Add find keyword feature
Users can add, delete, mark, etc. the current tasks.

However, if there are too many tasks, the user will not be able to
find the task that they are searching for.

The find feature allows the user to search for keywords and obtain the
tasks which have the keyword in their description.

This is done by adding find command (class FindCommand) and filter
method in the TaskList class in order to obtain the TaskList that has
been filtered by the keyword.