Update coding standard
Some methods were not named appropriately. Imports did not have the correct conventions.

Methods in Ui class are now named more appropriately according to the Java coding standard so that their actions are clearer by name.

Imports at the beginning of file are sorted in alphabetical order and spaced out. Doing so makes it easier to browse through the list of imports especially when there are many.