Fix save bug for deadline and event tasks.
Deadline and Event tasks are now save with their date attributes according to the read format of ""dd/mm/yyyy hhmm". To achieve this, the Task class has been made abstract with the save() abstract method for child classes to override with their own implementations of how it should be saved.

A minor change was also made to the DataHandler class to remove the wildcard import. Redundant imports were also removed from the State class and Sunday class.