Update deadline parser to recognize dates
Deadline task's deadlines are stored exclusively as strings.

Program is unable to make sense of the timing of deadlines.

Let's update the parser to detect deadline tasks created with
a deadline following the format "dd/mm/yyyy" and store deadlines
as dates.

Dates can be expressed more intuitively by the system. Features that
compare the deadlines of tasks can be added in the future.