Add GUI to the program
The program is using the terminal to print out the outputs.

Printing of the outputs does not look like a chatbot. Using an
application will make it look like a chatbot.

Let's add the classes for GUI and fxml files to use to build the app.

More efficient to use fxml file. Main class and Launcher class for easy
differentiation.