Create GUI for chat bot
Previously, there was no GUI and user had to interact with the bot
via CLI.

Create a GUI would improve the aesthetics of the application.

Let's use JavaFX to create a GUI, and for all Command classes, return
the String output to be processed instead of using System.out.println.