Improve user experience.
GUI only shows the first 5 tasks on command 'list' and abruptly closes the window on commnd 'bye'.

This provides the user bad with a bad experience.

Let's allow the dialogue box to expand as needed and put a 5 second delay after 'bye' command is entered to allow users to read the response properly.