Restructure code for GUI, using FXML
Previously, code for GUI is contained within the Duke class, with
the attributes being customised within a single method. This is
not scalable when application becomes larger as the method would
become too long.

Let's factor out the code into new classes and keep information of the
controls in FXML files instead. Then, use SceneBuilder to allow further
customisation such as CSS rules.

This makes it easier to expand the scale of the application when more
UI elements are added and abstract away the UI details from the Duke
class. Also, customisation is simpler.