Make parsing a queued service
If the user enters input fast enought, and some of them are invalid
inputs, the invalid input error message will race to be displayed with
the valid input result message. This may result in messages to be
displayed in the wrong order.

Let's make parsing a service that will be queued up to be executed in
the order that they are queued.

This synchronises the displyaing of command line result messages and
error messages from parsing.