Interpreter class: remove dead code
Interpreter class have some commented dead code which could never be reached.

The out-dated dead code decreases the readability of the code.

Deleting the dead code could follow the Code Quality better. This also allows the error not to be caught at the interpreter class, instead, it will be passed to execute() function Duke.java.

Let’s not handle the errors raised in the interprete() stage, instead, we handle it in the execute() function in Duke.java. This allows the errorMessage to be passed to the GUI and displayed to the users.

Refer to this tutorial on Code Quality
https://nus-cs2103-ay2223s2.github.io/website/se-book-adapted/chapters/codeQuality.html#code-quality