chore: Refactor code to follow SLAP
Currently, the parser class does not follow the SLAP principles due to
the validation logic.

This hurts the maintanability and readability of the Parser class as the
validation logic is occasionally abstracted while some are not. The
validation ocassionally comes in the middle of parsing which is not an
intuitive logic.

All validation logic are abstracted out and performed right at the start
when parsing the command. This gives a standardised order that is
easy to follow for readability.

Also, validation errors are now properly displayed to users due instead
of a generic message.