Cli class: extract message border to constant
The message border for the CLI version of Duke is a magic literal and is
duplicated (one for the start of the message, and one for the end of the
message).

Let's extract the message border into a private constant within the Cli
class.

Since message border is unchanging, store it as a constant.

There is no need to access it outside of the class hence give it the
private access modifier.

By having the message border literal stored in a variable, there is no
longer a need to duplicate the border.