Make commands display messages
Commands may have different ways of displaying messages such as a popup.
Allowing the commands to display the messages instead allows them to
display the message in the form that they desire.

Let's pull up the displaying logic from AppManager into the commands by
passing them the IoManager as well for execution.

Passing the command the AppManager instance might result in cyclic
dependencies.