Added A-Assertions.
Add assertions to Ui and TaskList methods.

There are several methods in Ui and TaskList that cannot be negative. The TaskList should not be negative, as it will indicate an error otherwise and terminate the program. It is better to assert this instead of terminating the program and having the user run it again.