Help command: launch a separate window
When a help command is issued, Duke outputs a long response
containing all commands it recognises.

This is hard to read due to the text being longer than the
window length, and needing the user to scroll up and down to
find the command they want to execute.

Let's add a separate window to contain all the commands that Duke
recognises, and launch this window when a help command is issued.

TODO:
* Improve the GUI of the help window