Popup: use stage to show popup
Popup is an undecorated window. Reimplementing the decorations of this
window is difficult and error prone. Unlike a window, Popup will by
default always appear on top of all others even when unfocused.

Let's use a Stage to contain the popup display instead.

This saves many lines of code and sanity.