Schedule: use Schedule timeable to display
There is no way to customise how a schedule for is displayed.

Let's use ScheduleTimeable instead to display schedules.

ScheduleTimeable defines methods to customise the way a schedule is
displayed. ScheduleComponent and DaySchedule are responsible for putting
the schedules on screen only and not deciding how it should be
displayed. Thus, the tasks should not be passed to those classes.
Passing the tasks will also limit the schedule graphic classes to only
work with AquaTasks which reduce resusability and abstraction.