Fix search by Add Streams
When introducing varargs to search for tasks, two bugs were created:
1. Broken search results due to empty string that match .contains()
2. Duplicated results

To fix 1. streams is used to remove empty string from varargs

To fix 2. HashSet to ensure only unique item are added, converted to
ArrayList afterwards to maintain compatibility with method return type

Solution was implemented with the help of Stackoverflow as Streams hard