Command classes : Refactor and clean some part of the code
Command class and its subclasses have some deprecated methods, such as
execute(), which is no longer used. This causes code redundancy.

This commit removes these codes.

The subclasses of Command also contain some long methods over 30 lines.
This reduces code readability and code quality.

For the duplicate code that is used in multiple subclasses, this commit
extracts them into public static methods in the superclass to be used in
 the subclasses, so as to reduce redundancy.

For the non-duplicate long methods within the subclasses, this commit
extracts them into private util methods in the subclasses. This
conforms to the OOP principle, and increases readability.

At the same time, this commit creates a new jar file, duke.jar, in
build/libs. The mainClassName in build.gradle is updated to be
“dukes.engine.Launcher”.