Add assertions to verify internal state of duke.
No assertions are currently in place.

This is a potential source of bugs if we add a new command and forget to handle it as it fails silently. If fxml is changed or image is removed, some objects might be null.

Let's introduce some assertions to detect these sources of error.