Made user experience smoother
The GUI does not automatically close after the bye command is entered.
The size of the window is also too small to display the initial list of
supported commands.

The GUI will now automatically close after a 3 second delay. Also, the
GUI window is a bit bigger to accommodate long lists of text in the
future if needed.