1:Add assertion: assert to command in main duke class
Creates assert on command to prevent receiving null

To prevent it from getting null as input for command so that the method will end and return an error message before continuing to use the method in the command class. As assertion acts as handling exception also.

Let’s verify if command is null before executing its methods first

2:Add command: fix bug of entering the task without an item

User enters task types as input is accepted even though there is no item afterwards

Resolving this issue by adding warning output for users is better as it allows them to follow the input requirements set by the duke in order for the command to be write into the system correctly. Failure to do so it will return an alert message asking the users for input, until users enter task in correct format: [task type + task item] or exit the app.

Let’s change the return type of addItem method to string so that it will return a message upon entering an incorrect input