Add C-Undo and Version Control.
A manual process is required to undo a command that
was accidentally entered in. For example, if the
user adds a task, and decides he wants to undo it,
he/she would need to manually enter a delete
command and specifies the task to be deleted. The
same can be said for mark and unmark commands.

With undo, this entire process can be reduced to a
single keyword, "undo". The undo operation is enabled
by keeping track of the last 10 changes. When the
undo command is entered, the most recent changed
state is popped and replaces the current state.

With these changes, it allows for version control to
take place as well. User can now checkout to a
specified version number and save accordingly.

Let's implement the undo function to improve the
user experience and enhance the quality of the application.