Refactor Parser
Parser uses nested conditionals and there's a lot of repeated code.

This makes readability difficult, and adding new commands troublesome.

Replacing the conditionals with a switch statement will allow the code to be more readable. Abstracting out many of the behaviors of the commands will also help to reduce code duplication.

Let's,
* Refactor the conditional statement into a switch statement using the Command enum
* Abstract out the behavior to extract delimiters from the parameter string into its own private function