Add assertions.
May not have an easy way for the coder to check that program behaviour is to be expected. Bugs may not be as easy to find.

As we continue to grow this Duke project, we need an easier way for programmers to detect bugs or areas where it is not performing as desired. This is especially when the code base is continually being modified.

It is done this way because it is simple. Assertions also stops the program flow when a bug is detected so as to not cause further complications.

Refer to CS2103T Textbook on assertions to see why assertions are important