Update GUI
The current GUI has no application header and icon.

The GUI is currently ugly with the default Java icon and header.

Let's add the GUI header and icon by using scene Image and some String
name called Duke in the Main class start method.

Best way to change the Icon and header of the application to look better
.