Refactor Sam's dialogs into an enum
Sam's dialogs are scattered everywhere as String literals.

Having the dialogs in a centralized position will make them
easier to keep track of.

Let's add a Dialog enum under Ui where all of Sam's dialogs
will be defined.

The Dialog enum is put under UI since dialogs
are related to the UI (as feedback for the user)