Add support for archiving task list
The user may have completed all tasks in the task list but do not want
to lose the data on the task by deleting all of them.

Let's
* add an ArchiveCommand class that handles archiving the user's task
  list to a file and clears the task list
* change Parser class to run ArchiveCommand::run(String, TaskList)
  whenever the user's input begins with "archive"

By encapsulating the logic for archiving the task list into the
ArchiveCommand, we can make it easier for developers to find said logic.
It also allows complies with the single-responsibility principle.