Use assert to document important assumptions
Some classes assume that code run previously has run correctly.

The program will fail to exit gracefully if these assumptions are not
true, and will hinder the usage of the application.

Let's add assert statements to Duke to ensure that the file path
provided points to a valid file, and to TaskList to ensure that the
commands provided to the constructor are not empty, since the code
assumes that the argument provided is not empty even if it is an
invalid command.

Assert statements ensure that the program runs correctly as expected.