Added jar files
The previous jar files are not updated to the master branch. There are
also some minor code quality improvements to be made, like unused
imports.

The jar file is built with meta-information within the resource folder.
The unused imports are fixed by removing the java import
lines in Main.java.