Critically examine code and refactor to improve the code quality
Some "happy path"s are unclear and are hard to follow, and some
JavaDoc comments are incorrectly formatted.

The expected flow of the code is currently difficult to follow, and
might cause problems in the future when the code has to be revisited
as a lot of time will have to be spent on figuring out what the
expected flow of the code is.

Let's:
- remove the finally block in Duke and add it to the try block
instead, which simplifies the error catching
- remove magic strings in Parser
- refactor the code fragment in Parser to validate whether the regex
is valid for the given input string to a separate method instead
- use fallthrough comments to simplify error catching in Parser
- rename confusing variable in TaskList to better represent the
variable's intended purpose
- remove unused imports in Storage and Ui
- fix JavaDoc comments in all files

Guard clauses and fallthrough blocks in switch statements can make the
"happy path" that the code is supposed to take when all is well easier
to read.