Implement A-Assertions
Various section of the code makes assumptions of the state of the program.

Add assert statements to act as null-checks and verify the
state of the program at various key areas.

Assert statements are disabled by default and will not affect the performance of the program.
However, it will indicate to the programmer the requirements of the state at that point.