Implement A-CodeQuality
Duke class initialises, styles the GUI and manages the user input which leads to
a very lengthy class that is difficult to maintain.

Extract the main gui portion into another class ChatPane which handles the
initialising and styling of the GUI.

Pulling it into its own class follows better OOP practise as encapsulation and abstraction
are utilised to make the GUI more maintainable. ChatPane inheriting the AnchorPane class allows
us to utilise a ChatPane instance like an AnchorPane.