A-FullCommitMessage
Current Duke prints output in the terminal but we want a GUI. There are
several additional methods added several classes to handle
returning a string instead of using System.out.println(). This is
because the GUI needs a String.

Added a GUI to the duke application.

Every class is part of a package to make things neater.