Add assert statement to document important assumptions
Components of class instance are assumed to exist prior to usage.

Such assumptions may not hold true, since code changes may
unknowingly remove such components.

Assert statements helps to ensure that important assumptions
holds at various points of the code.

Let's add assert statements to ensure that components of the class
instance are not null before interaction with that component.

Using assert statements is preferable since the programmer is
responsible for the validity of the components.