Refactor the parseRawCommand method in Parser
Previously, all parsing and checking of validity was done in the
parseRawCommand method. This means that the method is very long and
contains too much logic.

Let's extract out the logic for validating each type of command into
its own method, and call the corresponding validation method depending
on which command it is.

This simplifies the parseRawCommand method, and makes code neater as
it is easier to find the logic that validates each type of command,
and the method will not grow much larger whenever we add a new command
to the bot.