Add assertions
Add assertions

Assertions are used to define assumptions about the program state so that the runtime can verify them.

Asserts are added mainly whenever users perform an action that requires change to a task.

This was done to ensure that the modification of the task is done correctly and no error occurs.