Improve code quality: extract formatting of dates
Deadline and Events take in strings as arguments and have to convert
them to valid dates within their constructors.

Deadlines and Events should take in valid dates in the first place
without the conversion of strings.

Let's,
* extract the formatting of dates to be done in the parser class, where
  the DukeExceptions will be thrown if the date entered by user is
  invalid.
* create methods to convert string dates to LocalDateTime and LocalDate
  to reduce code duplication.

This way, DateTimeFormatters can be in one place without being in many
classes. Code will also be more readable and maintainable.