Add java assertions
There currently no way to display any assumptions for the code.

This should be changed to better illustrate the underlying assumptinos
in the code.

This problem is fixed by adding assertions to classes like storage,
parser and TaskList, which helps to show that some lines of code will
not be reached.