Implement support for Undo command.
There is no way to undo a previous command should the user makes a
mistake in calling a command that was unintended.

To provide a better experience, i.e. allow the user to rectify the
mistake, this commit implements support for the Undo command.

A stack is created to keep track of the previous commands executed,
excluding commands that are not applicable (List, Find, Exit, Undo).

The previous command can be obtained by popping the stack.

For relevant commands (Add, Delete, Mark, Unmark), an undo method is
implemented to undo the respective effect from executing those commands.