Add Java asserts
No use of assert statements in program code

Insufficient testing for bugs caused by the programmer in the code.

Let's add Java asserts to define assumptions in the program code, like
making sure the Command is not null.

Using Java asserts allows the runtime to verify these assumptions and
help find bugs in the code.