Abstract 'ParsedLoadedTask' into its own file
Move and renamed 'ParsedLoadedTask' to its own file 'ParsedTask',
which stores the relevant information of a task before its actual creation.
Currently, it is only used by 'Storage' when loading tasks from a saved file,
but should be later generalised to parse user commands as well.
(as opposed to current method of storing parsed info in an array)