Improve code quality and SlAP the commands.
There are several process command methods in the command classes that did
not apply SLAP as well as possible.

By abstracting out more complexity from the methods, other users will be able to read it with lesser ambiguity.

Complexity is handled by the ui and parser classes and the command classes will be a level of abstraction higher.