refactor: update `isSingleInputValid` to take in input with spaces
Refactor: include an additional parameter `isMulti` for the method `isSingleInputValid` to indicate if the input consists of multiple words delimited by `space` characters.

Method was only able to take inputs of one word which causes `find` commands to be inaccurately judged to be invalid.

By including the extra parameter, it can be used for all the commands that take in a single input, such as `done`, `undone`, `find`, `delete`.

Let's:
- utilise `isSingleInputValid` for all commands that take in a single input, even if the input might be a sentence made up of words delimited by `space` characters