Add assertions to the "duke", "storage" and "commands" packages
There are some classes such as Parser and TaskList that require
assertions to perform checks on variables.

For instance, when we edit the TaskList, such as when deleting,  we must ensure that a TaskList
has a size > 0. This ensures that we do not overdelete.

In Parser, we want to verify that the "bye" command must have a
.isBye() that returns true,  such that the while loop in Duke can be
closed.