Add assertions to Un/MarkCommand,Parser,DeadlineTask and CommandTask classes
Un/MarkCommand assertions checks that the tasklist size will never
be a negative number.Parser assertions checks that the parser
does not read an empty line as user input.DeadlineTask assertions
checks that the date and time provided conform to the necessary
format.

These ensure that if something were to go wrong it would be the
result of the user input instead of a problem with the program
itself.