Add reminder commands to the Nerdbot.
Nerdbot currently cannot send reminders to the user on upcoming deadlines and events.

This feature will add reminders for the user on upcoming deadlines based on the range of days from the current day.

By adding on to the existing command list, we can consistently update Nerdbot with new features easily, without changing any main structure of the program.