Duke: Add welcoming message.
When the user first download the program, it won't show welcoming
message to the screen and left it blank.

Need to print welcoming message to the screen otherwise user might think
that the program goes wrong.

Add a warm welcoming message.

Let's update the duke program so that the welcoming message can be shown
in GUI.