Add Assertions to parts of code
Assertions are needed to check that the value persisted downwards
into other deeper functions are not wrong

Assertions to add:
- index value - check that index input to the commands:
*MarkCommand
*UnmarkCommand
Tasklist.get__ commands

Index values are 0-indexed usually, but the user input index starts from the
convention from 1. We want to make sure the value is not equal to 0 by the time
it reaches the deeper nested functions, which account for the 1-index by subtracting
the received index by 1.