Improve code quality
There are some methods not using assertions or in good coding
style.

We use streams operations to replace some for loop to make the
code neater. Further, we replace magic strings and numbers with
constants to improve readability. Assertions are added to some
places to check program flow at runtime.