Improve code quality
Within certain file, there are method with too long implementation and
weak SLAP.

Code needs to be read, understood, and modified by other developers
later on. Even if the author do not intend to pass the code to someone
else, code quality is still important because the author will become a
'stranger' to his/her own code someday.

Use refactoring and guard clause to avoid improve readability.

Examine the code and refactor to improve the code quality where
necessary.